ECO-798: fix circular dependencies and compiling error.
