+++ packages/sdk/src/lib/index.ts
@@ -5,7 +5,6 @@ import * as Serialization from './Serialization';
 import * as Signer from './Signer';
 export * from './CLValue';
 export * from './RuntimeArgs';
-export * from './uref';
 export * from './CasperClient';
 
 export { Contracts, Keys, Serialization, DeployUtil, Signer };
